+++ src/__playground__/worker.ts
@@ -1,4 +1,3 @@
-import { threadId } from 'worker_threads';
 import { define, SharedMap } from '../index.js';
 import type { SharedMapTransferData } from '../index.js';
 
@@ -6,6 +5,6 @@ export const api = await define({
     async handleSharedMap(pair: SharedMapTransferData<{ foo: string; fizz: string; buzz: string }>) {
         const map = new SharedMap(pair);
 
-        console.log(await map.get('buzz'));
+        await map.set('buzz', 'xyzxyz');
     },
 });