+++ src/types/shared_map.ts
@@ -1,8 +1,8 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export type SharedMapTransferData<Data = Record<string, any>> = {
-    keys: Uint8Array;
-    values: Uint8Array;
-    status: Uint8Array;
+    __keys: Uint8Array;
+    __values: Uint8Array;
+    __status: Uint8Array;
 };
 
 export type KeyData = {