+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `TaskWorkerOptions` and `ServiceWorkerOptions` types to be exported as `LaunchTaskOptions` and `LaunchServiceOptions` instead.
 - Main file to export types `SharedArrayPair` and `MessengerTransferData`.
 - `SharedArrayPair` type name to `SharedMapTransferData`.
+- `Messenger.transfer()` to be a getter instead of a function.
 
 ### Added